Make message_creators module a subpackage
