+++ src/shared/line/message_creators/base_creators.py
@@ -1,0 +1,9 @@
+import abc
+
+from linebot import models
+
+
+class LineMessageCreator(abc.ABC):
+    @abc.abstractmethod
+    def generate(self) -> models.Message:
+        pass
